+++ src/services/Campaigns.services.js
@@ -1,13 +1,11 @@
 import axios from 'axios';
 import { campaignApi } from '../const/api';
 
-// export const RegisCampaign = async (Campaign) => {
-//   console.log(Campaign);
-//   await axios.post(`${api}addCampaign`, Campaign);
-//   alert('campagin successfully stored');
-// };
-
 export const RegisCampaign = async (Campaign) => {
+  if (!Campaign) {
+    alert('Campaign  was not successfully registered');
+    return;
+  }
   try {
     await axios.post(`${campaignApi}ADD`, Campaign);
     alert('campagin successfully stored');
@@ -17,6 +15,10 @@ export const RegisCampaign = async (Campaign) => {
 };
 
 export const getPersonalCampaigns = async (Email) => {
+  if (!Email) {
+    alert('Getting Personal Campaigns was not successful');
+    return;
+  }
   try {
     let results = await fetch(`${campaignApi}GET/${Email}`);
     let campaign = await results.json();
@@ -27,6 +29,10 @@ export const getPersonalCampaigns = async (Email) => {
 };
 
 export const delCampaign = async (ID) => {
+  if (!ID) {
+    alert('Campaign  was not successfully Deleted');
+    return;
+  }
   try {
     let endpoint = `${campaignApi}DELETE/${ID}`;
     await axios.delete(endpoint);
@@ -36,20 +42,43 @@ export const delCampaign = async (ID) => {
   }
 };
 
-// export const getCampaignID = async (website) => {
-//   try {
-//     let results = await fetch(`${campaignApi}GETID/${website}`);
-//     let campaign = await results.json();
-//     return campaign;
-//   } catch (error) {
-//     console.error(error);
-//   }
-// };
-
 export const UpdaCampaign = async (UpdatedCampaign) => {
+  if (!UpdatedCampaign) {
+    alert('Campaign  was not successfully Updated');
+    return;
+  }
+
   try {
     await axios.post(`${campaignApi}Update`, UpdatedCampaign);
     alert('campagin successfully updated');
+  } catch (error) {
+    console.error(error);
+  }
+};
+
+export const getCampaignID = async (ProductID) => {
+  if (!ProductID) {
+    alert('Getting  Campaigns ID was not successful');
+    return;
+  }
+  try {
+    let results = await fetch(`${campaignApi}GETCAMPAIGNID/${ProductID}`);
+    let campaign = await results.json();
+    return campaign;
+  } catch (error) {
+    console.error(error);
+  }
+};
+
+export const postCampaignDonation = async (CampaignID, unitPrice) => {
+  if (!CampaignID || !unitPrice) {
+    alert('Campaign Donation Amount was not successfully Updated');
+    return;
+  }
+
+  try {
+    await axios.post(`${campaignApi}DONATIONAMOUNT/${CampaignID}/${unitPrice}`);
+    alert('Donation to campaign was successfully updated');
   } catch (error) {
     console.error(error);
   }
